+++ src/base/SbCylinder.cpp
@@ -222,7 +222,7 @@ SbCylinder::intersect(const SbLine& l, SbVec3f& enter, SbVec3f& exit) const
   // this distance is larger than the radius of the cylinder, there's
   // of course no intersection.
   n.normalize();
-  float d = fabs(cv.dot(n));
+  float d = (float)fabs(cv.dot(n));
   if(d > this->radius) return FALSE;
 
   // There's an intersection, now find the parameter for the plane
@@ -232,8 +232,8 @@ SbCylinder::intersect(const SbLine& l, SbVec3f& enter, SbVec3f& exit) const
   tmp = n.cross(this->axis.getDirection());
   tmp.normalize();
 
-  float s = fabs(sqrt(this->radius*this->radius - d*d) /
-                 l.getDirection().dot(tmp));
+  float s = (float)fabs(sqrt(this->radius*this->radius - d*d) /
+                        l.getDirection().dot(tmp));
 
   float enterparam = t - s;
   float exitparam = t + s;
