+++ src/base/SbRotation.cpp
@@ -163,8 +163,8 @@ void
 SbRotation::getValue(SbVec3f& axis, float& radians) const
 {
   if((this->quat[3] >= -1.0f) && (this->quat[3] <= 1.0f)) {
-    radians = acos(this->quat[3]) * 2.0f;
-    float scale = sin(radians / 2.0f);
+    radians = float(acos(this->quat[3])) * 2.0f;
+    float scale = (float)sin(radians / 2.0f);
 
     if(scale != 0.0f) {
       axis[0] = this->quat[0] / scale;
@@ -297,7 +297,7 @@ SbRotation::setValue(const SbMatrix& m)
   float scalerow = m[0][0] + m[1][1] + m[2][2];
 
   if (scalerow > 0.0f) {
-    float s = sqrt(scalerow + 1.0f);
+    float s = (float)sqrt(scalerow + 1.0f);
     this->quat[3] = s * 0.5f;
     s = 0.5f / s;
 
@@ -313,7 +313,7 @@ SbRotation::setValue(const SbMatrix& m)
     int j = (i+1)%3;
     int k = (j+1)%3;
 
-    float s = sqrt((m[i][i] - (m[j][j] + m[k][k])) + 1.0f);
+    float s = (float)sqrt((m[i][i] - (m[j][j] + m[k][k])) + 1.0f);
 
     this->quat[i] = s * 0.5f;
     if(s != 0.0f) s = 0.5f / s;
@@ -345,9 +345,9 @@ SbRotation::setValue(const SbVec3f& axis, const float radians)
 
   // From <http://www.automation.hut.fi/~jaro/thesis/hyper/node9.html>.
 
-  this->quat[3] = cos(radians/2);
+  this->quat[3] = (float)cos(radians/2);
 
-  const float sineval = sin(radians/2);
+  const float sineval = (float)sin(radians/2);
   SbVec3f a = axis;
   a.normalize();
   this->quat[0] = a[0] * sineval;
@@ -392,7 +392,7 @@ SbRotation::setValue(const SbVec3f& rotateFrom, const SbVec3f& rotateTo)
 
   // FIXME: this should perhaps be "centralized" somewhere. 19990401 mortene.
   // Make sure we can take the cross-product without underflow.
-  const float tolerance = sqrt(sqrt(FLT_MIN));
+  const float tolerance = (float)sqrt(sqrt(FLT_MIN));
 
   // Check if they are parallel and pointing in the same direction.
   if (fabs(dot - 1.0f) < tolerance) {
@@ -416,10 +416,10 @@ SbRotation::setValue(const SbVec3f& rotateFrom, const SbVec3f& rotateTo)
     t.normalize();
     // The fabs() wrapping is to avoid problems when `dot' "overflows"
     // a tiny wee bit, which can lead to sqrt() returning NaN.
-    t *= sqrt(0.5f * fabs(1.0f - dot));
+    t *= (float)sqrt(0.5f * fabs(1.0f - dot));
     // The fabs() wrapping is to avoid problems when `dot' "underflows"
     // a tiny wee bit, which can lead to sqrt() returning NaN.
-    this->setValue(t[0], t[1], t[2], sqrt(0.5f * fabs(1.0f + dot)));
+    this->setValue(t[0], t[1], t[2], (float)sqrt(0.5 * fabs(1.0 + dot)));
   }
 
   return *this;
@@ -567,10 +567,10 @@ SbRotation::slerp(const SbRotation& rot0, const SbRotation& rot1, float t)
   // scale factors.
   if((1.0 - dot) > 0.0f) {
 #if 1 // Spherical interpolation (geometrically correct).
-    float angle = acos(dot);
-    float sinangle = sin(angle);
-    float scale0 = sin((1.0 - t) * angle) / sinangle;
-    float scale1 = sin(t * angle) / sinangle;
+    float angle = (float)acos(dot);
+    float sinangle = (float)sin(angle);
+    float scale0 = float(sin((1.0 - t) * angle)) / sinangle;
+    float scale1 = float(sin(t * angle)) / sinangle;
 #else // Linear interpolation (faster).
     float scale0 = 1.0 - t;
     float scale1 = t;
